Renamed files and made additional changes based on feedback.
File names have been changed to make it more clear which file is which. Additionally, some empty props have been removed, and certain interfaces, such as IUser and IPassword settings, have been moved to their own files. Users can no longer change their date of birth or name.
