+++ Client/src/Components/Profile/ProfileForm.tsx
@@ -3,21 +3,10 @@ import * as Yup from 'yup'
 import { Button, Container } from '@material-ui/core'
 import { Form, Formik } from 'formik'
 
-import { ProfileSection } from './ProfileSection/ProfileSection'
+import { IUser } from '../../Models/Profile/IProfileModel'
 import React from 'react'
+import { UserForm } from './UserDetailSection/UserForm'
 
-export interface IUser {
-  name: string,
-  email: string,
-  dateOfBirth: string,
-  organization: string,
-  password: string
-}
-
-export interface IPasswordSettings {
-  password: string,
-  passwordConfirmation: string
-}
 interface IProps {
   user: IUser,
   onSubmit: (user: IUser) => void
@@ -28,9 +17,7 @@ export default function ProfileForm(props: IProps) {
   const { user, onSubmit } = props
 
   const validationSchema = Yup.object().shape({
-    name: Yup.string().required(),
     email: Yup.string().email('Please enter a valid email').required(),
-    dateOfBirth: Yup.string().required(),
     organization: Yup.string().required()
   })
 
@@ -42,7 +29,7 @@ export default function ProfileForm(props: IProps) {
         onSubmit={onSubmit}
       >
         <Form>
-          <ProfileSection />
+          <UserForm />
           <Button variant="contained" color="primary" type="submit">Update User</Button>
         </Form>
       </Formik>
