+++ Client/src/Models/Profile/IProfileModel.ts
@@ -1,0 +1,12 @@
+export interface IUser {
+  name: string,
+  email: string,
+  dateOfBirth: string,
+  organization: string,
+  password: string
+}
+
+export interface IPasswordSettings {
+  password: string,
+  passwordConfirmation: string
+}
